replaced the CLI wrapping in favor of a web API. Packed all the tasks into a single consumer. Tested and working
