+++ config.js
@@ -1,10 +1,13 @@
 const config = {
     saimeUrl: 'https://tramites.saime.gob.ve/index.php?r=usuario/usuario/BuscarSaimeContacto',
     cneUrlBase: 'http://www.cne.gov.ve/web/registro_electoral/ce.php',
-    mongoUrl: 'mongodb://localhost:27017',
+    mongoUrl: `mongodb://${process.env.MONGO_HOST}`,
     saimeCollection: 'saime',
     cneCollection: 'cne',
     mongoDB: 'venez',
+    redisUrl: `redis://${process.env.REDIS_HOST}`,
+    apiHost: '0.0.0.0',
+    apiPort: 3333,
 }
 
 module.exports = config;
